Fix BBOX filtering of all OSM changeset feature
When using "geogig osm apply-diff", ALL changes were "filter out" (removed)
because the default BBOX was empty.  Since the OSM dataset is EPSG:4326, I
set it to the World Bounds.

Related, I also suppressed the message showing the boundles check so
it only reports "bad" data.  It was too difficult to see problems
in a sea of reports (one problem in a millions of lines of text).

Signed-off-by: DBlasby <dblasby@boundlessgeo.com>